+++ data-effects-core/src/Control/Effect/Key.hs
@@ -14,7 +14,7 @@ Portability :  portable
 module Control.Effect.Key where
 
 import Control.Applicative (Alternative)
-import Control.Effect (SendIns (sendIns), SendSig (sendSig))
+import Control.Effect (SendFOE (sendFOE), SendHOE (sendHOE))
 import Control.Monad (MonadPlus)
 import Control.Monad.Except (MonadError)
 import Control.Monad.Fix (MonadFix)
@@ -24,15 +24,15 @@ import Control.Monad.Reader (MonadReader)
 import Control.Monad.State (MonadState)
 import Control.Monad.Writer (MonadWriter)
 import Data.Coerce (coerce)
-import Data.Effect (InsClass, SigClass)
+import Data.Effect (EffectF, EffectH)
 import Data.Effect.HFunctor (HFunctor, hfmap)
 import Data.Kind (Type)
 
-class SendInsBy key (ins :: InsClass) f | key f -> ins where
-    sendInsBy :: ins a -> f a
+class SendFOEBy key (ins :: EffectF) f | key f -> ins where
+    sendFOEBy :: ins a -> f a
 
-class SendSigBy key (sig :: SigClass) f | key f -> sig where
-    sendSigBy :: sig f a -> f a
+class SendHOEBy key (sig :: EffectH) f | key f -> sig where
+    sendHOEBy :: sig f a -> f a
 
 -- | A wrapper data type to represent sending an effect to the carrier @f@ with the specified key.
 newtype ByKey key (f :: Type -> Type) a = ByKey {runByKey :: f a}
@@ -57,10 +57,10 @@ key :: forall key f a. ByKey key f a -> f a
 key = runByKey
 {-# INLINE key #-}
 
-instance SendInsBy key ins f => SendIns ins (ByKey key f) where
-    sendIns = ByKey . sendInsBy @key
-    {-# INLINE sendIns #-}
+instance (SendFOEBy key ins f) => SendFOE ins (ByKey key f) where
+    sendFOE = ByKey . sendFOEBy @key
+    {-# INLINE sendFOE #-}
 
-instance (SendSigBy key sig f, HFunctor sig) => SendSig sig (ByKey key f) where
-    sendSig = ByKey . sendSigBy @key . hfmap coerce
-    {-# INLINE sendSig #-}
+instance (SendHOEBy key sig f, HFunctor sig) => SendHOE sig (ByKey key f) where
+    sendHOE = ByKey . sendHOEBy @key . hfmap coerce
+    {-# INLINE sendHOE #-}