+++ data-effects/src/Data/Effect/Provider/Implicit.hs
@@ -16,30 +16,30 @@ Portability :  portable
 module Data.Effect.Provider.Implicit where
 
 data ImplicitProvider' c i e (f :: Type -> Type) (a :: Type) where
-    WithImplicit ::
-        i ->
-        (forall g. (c g, e g) => (forall x. f x -> g x) -> g a) ->
-        ImplicitProvider' c i e f a
+    WithImplicit
+        :: i
+        -> (forall g. (c g, e g) => (forall x. f x -> g x) -> g a)
+        -> ImplicitProvider' c i e f a
 
 makeKeyedEffect [] [''ImplicitProvider']
 
 type MonadImplicitProvider i e = ImplicitProvider Monad i e
 type ApplicativeImplicitProvider i e = ImplicitProvider Applicative i e
 
-(.!) ::
-    forall c e i f a.
-    SendSigBy ImplicitProviderKey (ImplicitProvider' c i e) f =>
-    i ->
-    (forall g. (c g, e g) => g a) ->
-    f a
+(.!)
+    :: forall c e i f a
+     . (SendHOEBy ImplicitProviderKey (ImplicitProvider' c i e) f)
+    => i
+    -> (forall g. (c g, e g) => g a)
+    -> f a
 i .! m = withImplicit i \_ -> m
 {-# INLINE (.!) #-}
 
-(..!) ::
-    forall c e i f a.
-    SendSigBy ImplicitProviderKey (ImplicitProvider' c i e) f =>
-    i ->
-    (forall g. (c g, e g) => f ~> g -> g a) ->
-    f a
+(..!)
+    :: forall c e i f a
+     . (SendHOEBy ImplicitProviderKey (ImplicitProvider' c i e) f)
+    => i
+    -> (forall g. (c g, e g) => f ~> g -> g a)
+    -> f a
 i ..! f = withImplicit i f
 {-# INLINE (..!) #-}