+++ data-effects/src/Prelude.hs
@@ -16,7 +16,7 @@ module Prelude (
 ) where
 
 import Control.Effect (type (<:), type (<<:), type (~>))
-import Control.Effect.Key (SendInsBy, SendSigBy)
+import Control.Effect.Key (SendFOEBy, SendHOEBy)
 import Data.Default (Default (def))
 import Data.Effect.Key.TH (makeKeyedEffect)
 import Data.Effect.TH (makeEffect, makeEffect', makeEffectF, makeEffectH)