+++ data-effects-th/ChangeLog.md
@@ -5,7 +5,7 @@
 * Initial public release.
 
 ## 0.1.1.0 -- 2024-09-01
-* Fix an issue where constraints in the effect constructor were not properly reflected in 'genLiftInsPatternSynonyms'.
+* Fix an issue where constraints in the effect constructor were not properly reflected in 'genLiftFOEPatternSynonyms'.
 * Fixed an issue where the auto-generation would fail when the name of the effect constructor started with a colon (:) operator.
 
 ## 0.1.2.1 -- 2024-09-26